+++ src/models/GQLValue.ts
@@ -1,5 +1,57 @@
-import Record from 'dataclass';
+import {List} from 'immutable';
 
-export class GQLValue<T> extends Record<GQLValue<T>> {
+interface IGQLValue {
+  value: any;
+}
+
+export class GQLValue implements IGQLValue {
   public value: any;
+
+  constructor(value: any) {
+    this.value = value;
+  }
+}
+
+export class GQLBooleanValue extends GQLValue {
+    public value: boolean;
+    constructor(boolValue: boolean) {
+        super();
+        this.value = boolValue;
+    }
+}
+
+export class GQLArrayValue extends GQLValue {
+    public value: List<any>;
+    constructor(arrValue: []) {
+        super();
+        this.value = List(arrValue);
+    }
+}
+
+export class GQLNumberValue extends GQLValue {
+    public value: number;
+    constructor(numValue: number) {
+      super();
+      this.value = numValue;
+    }
+}
+
+export class GQLDoubleValue extends GQLNumberValue {}
+export class GQLIntValue extends GQLNumberValue {}
+export class GQLLongValue extends GQLNumberValue {}
+
+export class GQLEnumValue extends GQLValue {
+    public value: string;
+    constructor(strValue: string) {
+        super();
+        this.value = strValue;
+    }
+}
+
+export class GQLStringValue extends GQLValue {
+    public value: string;
+    constructor(strValue: string) {
+        super();
+        this.value = strValue;
+    }
 }
